requirements.txt
anndata~=0.8.0
fastavro
google-api-python-client
google-cloud-bigquery
google-cloud-storage
numpy~=1.23.1
scipy
